+++ SecurityLib/Readers/DataReader.cs
@@ -44,8 +44,8 @@ namespace WolfInv.com.SecurityLib
         public abstract ExpectList<T> ReadHistory<T>(string begt, string endt) where T : TimeSerialData;
         public abstract ExpectList<T> ReadNewestData<T>(DateTime fromdate) where T : TimeSerialData;
         public abstract ExpectList<T> ReadNewestData<T>(int LastLng) where T : TimeSerialData;
-        public abstract ExpectList<T> ReadNewestData<T>(int ExpectNo, int Cnt) where T : TimeSerialData;
-        public abstract ExpectList<T> ReadNewestData<T>(int ExpectNo, int Cnt, bool FromHistoryTable) where T : TimeSerialData;
+        public abstract ExpectList<T> ReadNewestData<T>(long ExpectNo, int Cnt) where T : TimeSerialData;
+        public abstract ExpectList<T> ReadNewestData<T>(long ExpectNo, int Cnt, bool FromHistoryTable) where T : TimeSerialData;
         public abstract int SaveChances<T>(List<ChanceClass<T>> list, string strDataOwner=null) where T : TimeSerialData;
         public abstract int SaveHistoryData<T>(ExpectList<T> InData) where T : TimeSerialData;
         public abstract int SaveNewestData<T>(ExpectList<T> InData) where T : TimeSerialData;
